Fix Spatial Options Display (#1051)
* Fix Spatial Options Display

* Comment.
